Refactor GoldButtonClicker and VaultDisplay to improve type handling and UI updates, changing RefreshResourceUI parameter to long and correcting ratio calculation for fill bar display.

+++ Assets/Game/2Game/Script/Clicker/GoldButtonClicker.cs
@@ -46,7 +46,7 @@ public class GoldButtonClicker : MonoBehaviour
             GameManager.Instance.AddGold(value);
     }
 
-    void RefreshResourceUI(double _)
+    void RefreshResourceUI(long _)
     {
         if (GameManager.Instance == null) return;
